Nail reek version to the point it depends on ruby_parser-2
Reek 1.3 (the current release) depends on ruby_parser-3.
flay and flog could be updated, but for roodi there is no release that
supports ruby_parser-3 until a decision is done we cherry pick a version
of reek that does not force us to use ruby_parser-3.

IMHO we should drop roodi.