Don't require config files to be present initially
For most of the metric tools (all but yardstick),
it is enough to just require the config once the
tool is actually invoked. Yardstick however needs
to define its rake task upfront, therefore imposing
the need to have a default config available.

This is just a quick way to work around that for
yardstick, the change to tasks/metrics/flog.rake
goes inline with what all the other tools do.

All this is needed to provide a devtools:init task
that copies our default metric configs over to a
project's root directory.

That init task will be added in the next commit.