Move mutant rake task body to class
I removed the `zombify` logic because it doesn't seem to be
valid usage of mutant anymore.  I try to keep the code as unchanged
as possible when first extracting from rake tasks but in this case
I didn't want to test invalid behavior for the sake of having
a clean copy over.