Update rubocop to 0.44.1
This change adds a few disables for the new UnifiedInteger cop. Replacing
`Fixnum` with `Integer` would be semantically equivalent if `is_a?` checks
were being performed but the `Fixnum` usage in this project is checking
`instance_of?(Fixnum)` which will not be semantically equivalent to
`instance_of?(Integer)` until Ruby 2.4.